Enable color_threshold and passthorugh tests in hw

+++ reference_designs/ipu-xrt/vision_pipelines/color_threshold/test.cpp
@@ -165,28 +165,30 @@ int main(int argc, const char *argv[]) {
     if (srcVec[i] <= 50) { // Obviously change this back to 100
       if (*(bufOut + i) != 0) {
         if (errors < max_errors)
-          std::cout << "Error: " << (uint8_t) * (bufOut + i) << " at " << i
-                    << " should be zero "
-                    << " : input " << std::dec << (uint8_t)srcVec[i]
+          std::cout << "Error: " << (uint32_t)(uint8_t) * (bufOut + i) << " at "
+                    << i << " should be zero "
+                    << " : input " << std::dec << (uint32_t)(uint8_t)srcVec[i]
                     << std::endl;
         errors++;
       } else {
-        std::cout << "Below threshold:   " << (uint8_t) * (bufOut + i) << " at "
-                  << i << " is correct "
-                  << " : input " << std::dec << (uint8_t)srcVec[i] << std::endl;
+        std::cout << "Below threshold:   " << (uint32_t)(uint8_t) * (bufOut + i)
+                  << " at " << i << " is correct "
+                  << " : input " << std::dec << (uint32_t)(uint8_t)srcVec[i]
+                  << std::endl;
       }
     } else {
       if (*(bufOut + i) != UINT8_MAX) {
         if (errors < max_errors)
-          std::cout << "Error: " << (uint8_t) * (bufOut + i) << " at " << i
-                    << " should be UINT8_MAX "
-                    << " : input " << std::dec << (uint8_t)srcVec[i]
+          std::cout << "Error: " << (uint32_t)(uint8_t) * (bufOut + i) << " at "
+                    << i << " should be UINT8_MAX "
+                    << " : input " << std::dec << (uint32_t)(uint8_t)srcVec[i]
                     << std::endl;
         errors++;
       } else {
-        std::cout << "Above threshold:  " << (uint8_t) * (bufOut + i) << " at "
-                  << i << " is correct "
-                  << " : input " << std::dec << (uint8_t)srcVec[i] << std::endl;
+        std::cout << "Above threshold:  " << (uint32_t)(uint8_t) * (bufOut + i)
+                  << " at " << i << " is correct "
+                  << " : input " << std::dec << (uint32_t)(uint8_t)srcVec[i]
+                  << std::endl;
       }
     }
   }
